Base Early Exit version off of generic RexNext
